requirements.txt
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
flask==2.3.3
werkzeug==2.3.7 